* commit '7ffc155091d2ffa158f5866ffe8fa4588f62cf09':
  Fix issue when there is no JobHistoryErrors to clean up

+++ Source/kCura.IntegrationPoints.Management/Tasks/CleanUnlinkedJobHistoryErrorsTask.cs
@@ -25,9 +25,12 @@ namespace kCura.IntegrationPoints.Management.Tasks
 
 		private void CleanInWorkspace(int workspaceArtifactId)
 		{
-			var jobHistories = _unlinkedJobHistoryService.FindUnlinkedJobHistories(workspaceArtifactId);
+			List<int> jobHistories = _unlinkedJobHistoryService.FindUnlinkedJobHistories(workspaceArtifactId);
 
-			_deleteHistoryErrorService.DeleteErrorAssociatedWithHistories(jobHistories, workspaceArtifactId);
+			if (jobHistories.Count > 0)
+			{
+				_deleteHistoryErrorService.DeleteErrorAssociatedWithHistories(jobHistories, workspaceArtifactId);
+			}
 		}
 	}
 }
